jupyter runtimes into separate module
