+++ runtimelib/src/jupyter_runtime.rs
@@ -1,0 +1,49 @@
+use crate::jupyter_dirs;
+use serde::{Deserialize, Serialize};
+use serde_json::from_str;
+use tokio::fs;
+
+#[derive(Serialize, Clone)]
+pub struct JupyterEnvironment {
+    process: String,
+    argv: Vec<String>,
+    display_name: String,
+    language: String,
+}
+
+#[derive(Serialize, Deserialize, Clone, Debug)]
+pub struct JupyterRuntime {
+    pub shell_port: u16,
+    pub iopub_port: u16,
+    pub stdin_port: u16,
+    pub control_port: u16,
+    pub hb_port: u16,
+    pub kernel_name: String,
+    pub ip: String,
+    key: String,
+    pub transport: String, // TODO: Enumify with tcp, ipc
+    signature_scheme: String,
+    // We'll track the connection file path here as well
+    #[serde(skip_deserializing)]
+    pub connection_file: String,
+}
+
+pub async fn get_jupyter_runtime_instances() -> Vec<JupyterRuntime> {
+    let runtime_dir = jupyter_dirs::runtime_dir();
+    let mut runtimes = Vec::new();
+
+    if let Ok(mut entries) = fs::read_dir(runtime_dir).await {
+        while let Some(entry) = entries.next_entry().await.unwrap_or(None) {
+            let path = entry.path();
+            if path.is_file() && path.extension().and_then(|s| s.to_str()) == Some("json") {
+                let content = fs::read_to_string(&path).await.unwrap_or_default();
+                if let Ok(mut runtime) = from_str::<JupyterRuntime>(&content) {
+                    runtime.connection_file = path.to_str().unwrap_or_default().to_string();
+                    runtimes.push(runtime);
+                }
+            }
+        }
+    }
+
+    runtimes
+}
